# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pilot Web UI API",
    "description": "BFF server API for Pilot Web UI — chat sessions, platform channels, model management, skills, memory, logs, and terminal.",
    "version": "0.2.6"
  },
  "servers": [
    { "url": "http://localhost:8648", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/pilot/available-models": {
      "get": {
        "tags": ["Models"],
        "summary": "Fetch available models",
        "operationId": "getAvailableModels",
        "description": "Fetches models from all credential pool endpoints in ~/.pilot/auth.json",
        "responses": {
          "200": { "description": "Model list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AvailableModelsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to fetch", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/config/models": {
      "get": {
        "tags": ["Models"],
        "summary": "Get model config",
        "operationId": "getModelConfig",
        "responses": {
          "200": { "description": "Model config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelConfigResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/config/model": {
      "put": {
        "tags": ["Models"],
        "summary": "Update default model",
        "operationId": "updateDefaultModel",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateModelRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing default model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to update", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/config/providers": {
      "post": {
        "tags": ["Models"],
        "summary": "Add custom provider",
        "operationId": "addProvider",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddProviderRequest" } } } },
        "responses": {
          "200": { "description": "Added", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing required fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to add", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/config/providers/{poolKey}": {
      "delete": {
        "tags": ["Models"],
        "summary": "Delete provider",
        "operationId": "deleteProvider",
        "parameters": [{ "name": "poolKey", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Credential pool key (URL-encoded)" }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Cannot delete last provider", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Provider not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Failed to delete", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/skills": {
      "get": {
        "tags": ["Skills"],
        "summary": "List all skills",
        "operationId": "listSkills",
        "responses": {
          "200": { "description": "Skills grouped by category", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/skills/toggle": {
      "put": {
        "tags": ["Skills"],
        "summary": "Toggle skill enabled state",
        "operationId": "toggleSkill",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ToggleSkillRequest" } } } },
        "responses": {
          "200": { "description": "Toggled", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to toggle", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/skills/{category}/{skill}/files": {
      "get": {
        "tags": ["Skills"],
        "summary": "List skill files",
        "operationId": "listSkillFiles",
        "parameters": [
          { "name": "category", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "skill", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "File list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SkillFilesResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/skills/{path}": {
      "get": {
        "tags": ["Skills"],
        "summary": "Read skill file",
        "operationId": "readSkillFile",
        "parameters": [{ "name": "path", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Relative path under skills directory (e.g. category/skill/SKILL.md)" }],
        "responses": {
          "200": { "description": "File content", "content": { "application/json": { "schema": { "type": "object", "properties": { "content": { "type": "string" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "description": "Access denied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "File not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/memory": {
      "get": {
        "tags": ["Memory"],
        "summary": "Read memory files",
        "operationId": "getMemory",
        "responses": {
          "200": { "description": "Memory content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoryResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        },
        "security": [{ "BearerAuth": [] }]
      },
      "post": {
        "tags": ["Memory"],
        "summary": "Write memory file",
        "operationId": "saveMemory",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveMemoryRequest" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid section", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to save", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Auth token from dist/server/data/.token or AUTH_TOKEN env var"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "AvailableModelsResponse": {
        "type": "object",
        "properties": {
          "default": { "type": "string" },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "provider": { "type": "string" },
                "label": { "type": "string" },
                "base_url": { "type": "string" },
                "models": { "type": "array", "items": { "type": "string" } }
              }
            }
          }
        }
      },
      "ModelConfigResponse": {
        "type": "object",
        "properties": {
          "default": { "type": "string" },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "provider": { "type": "string" },
                "models": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": { "id": { "type": "string" }, "label": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      },
      "UpdateModelRequest": {
        "type": "object",
        "properties": {
          "default": { "type": "string", "description": "Model ID to set as default" },
          "provider": { "type": "string", "description": "Provider key (optional)" }
        },
        "required": ["default"]
      },
      "AddProviderRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Display name" },
          "base_url": { "type": "string", "description": "API base URL" },
          "api_key": { "type": "string", "description": "API key" },
          "model": { "type": "string", "description": "Model ID" },
          "providerKey": { "type": "string", "description": "Credential pool key (optional)" }
        },
        "required": ["name", "base_url", "api_key", "model"]
      },
      "SkillsResponse": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "description": { "type": "string" },
                "skills": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "description": { "type": "string" },
                      "enabled": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ToggleSkillRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Skill name" },
          "enabled": { "type": "boolean", "description": "Target enabled state" }
        },
        "required": ["name", "enabled"]
      },
      "SkillFilesResponse": {
        "type": "object",
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "string", "description": "Relative path within skill directory" },
                "name": { "type": "string", "description": "Filename" }
              }
            }
          }
        }
      },
      "MemoryResponse": {
        "type": "object",
        "properties": {
          "memory": { "type": "string", "description": "Contents of MEMORY.md" },
          "user": { "type": "string", "description": "Contents of USER.md" },
          "memory_mtime": { "type": "number", "nullable": true, "description": "Last modified timestamp (ms)" },
          "user_mtime": { "type": "number", "nullable": true, "description": "Last modified timestamp (ms)" }
        }
      },
      "SaveMemoryRequest": {
        "type": "object",
        "properties": {
          "section": { "type": "string", "enum": ["memory", "user"], "description": "Which memory file to write" },
          "content": { "type": "string", "description": "Full file content" }
        },
        "required": ["section", "content"]
      },
...
    }
  }
}